Add namespace prefix to headers
The types ksanecore provides are in the KSaneCore namespace. It is convention to add the namespace as a prefix to the headers the consumer includes

e.g. #include<Option> should be #include<KSaneCore/Option>

This avoids adding generic header names like Option to the global 'include namespace' and avoids potential conflicts with other libraries