Don't crash scanner apps (skanlite, skanpage) when searching for scanners
This is triggered eg. with network scanners where scanning takes longer time.
As comment says: this defeats the idea of using a scanner thread,
but the proper fix would be more elaborate (ie. do the rest of initialization
only when the thread is finished).
Still makes sense to have this hotfix, especially as this is an old bug...

BUG: 444987
BUG: 468053