Fix unselectable A4 page size
Problem: observed on a Brother MFC-L3770CDW using the program Skanpage. When
setting A4 as page size, the combobox switches automatically to the
"Custom" value.

Solution: I do not understand the exact purpose of the code but the logic of
the original code looked wrong and this change fixed the problem on my
system.

My scanner for A4 has some wiggleRoom for the width and nothing for the height,
switching the dimensions in the sum probably caused the problem. A
scanner with different geometry will probably behave differently.

Possible related bug:
BUG: 512876